Updates and upgrades CUDA support
- Kernel launch routines have new state Programming accelerator
- Fixes communication lines to show kernel scheduling - execution and memcpy partners
- CUPTI tracks unknown CUDA calls

+++ src/merger/paraver/opencl_prv_semantics.c
@@ -54,7 +54,7 @@ static int OpenCL_Acc_Call (event_t * event, unsigned long long time,
 {
 	unsigned state;
 	unsigned EvType, nEvType;
-	unsigned long long EvValue, nEvValue;
+	UINT64 EvValue, nEvValue;
 	
 	UNREFERENCED_PARAMETER(fset);
 
@@ -134,7 +134,7 @@ static int OpenCL_Host_Call (event_t * event, unsigned long long time,
 {
 	unsigned state;
 	unsigned EvType, nEvType;
-	unsigned long long EvValue, nEvValue;
+	UINT64 EvValue, nEvValue;
 	
 	UNREFERENCED_PARAMETER(fset);
